Revert "VPAAMP-157 Spurious PlaybackSpeedChanged event (#1341)"
This reverts commit aecaf802692dc659b394e35447746d26ee304da9.

+++ test/utests/tests/TsProcessorTests/sendSegmentTests.cpp
@@ -302,7 +302,7 @@ TEST_F(sendSegmentTests, CallsetBasePTSTest5)
 
 TEST_F(sendSegmentTests, CallsetBasePTSTest6)
 {
-    double position = static_cast<double>(LLONG_MAX);
+    double position = LLONG_MAX;
     long long pts = LLONG_MIN;
     mTSProcessor->CallsetBasePTS(position, pts);
 }
